+++ UI/Vdom.ts
@@ -20,10 +20,7 @@ import * as _MountMethodsCaller from './_vdom/Synchronizer/resources/MountMethod
  */
 export { default as Synchronizer } from './_vdom/Synchronizer/Synchronizer';
 export { Debug };
-export {
-   getChangedOptions,
-   createInstance
-} from './_vdom/Synchronizer/resources/DirtyChecking';
+export { createInstance } from './_vdom/Synchronizer/resources/ControlNode';
 export { DirtyChecking };
 export { DirtyCheckingCompatible };
 export { default as DOMEnvironment } from './_vdom/Synchronizer/resources/DOMEnvironment';
