+++ UI/_executor/_Markup/Vdom/IVdomType.ts
@@ -33,4 +33,4 @@ export interface IGeneratorControlNode {
 }
 
 // Тип для генераторов Vdom-нод
-export type GeneratorNode = IGeneratorVNode | IGeneratorControlNode;
+export type GeneratorNode = IGeneratorVNode & IGeneratorControlNode;
